requirements.txt
selenium==4.13.0
beautifulsoup4==4.12.2
Flask>=2.2.0
Flask-MySQL>=1.5.2
Flask-SQLAlchemy >=3.1.1
Flask-Migrate >= 4.1.0
PyMySQL
python-dotenv
# numpy
numpy
pandas
mysql-connector-python
scikit-learn
matplotlib
seaborn
nltk
joblib

